\name{AnalysisApp}
\alias{AnalysisApp}
\title{
Creation of a Shiny-based interface to perform an MCPMod analysis of a dose-finding trial
}
\description{
This function creates a Shiny-based graphical user interface to perform MCPMod-based analysis of a dose-finding trial.
}
\usage{
AnalysisApp()
}
\author{
Alex Dmitrienko <admitrienko@medianainc.com>
}
\seealso{
  \code{\link{MCPModAnalysis}}, \code{\link{AnalysisReport}}
}
\examples{
  \donttest{
# Launch a Shiny-based interface to perform an MCPMod-based analysis of a dose-finding trial
AnalysisApp()
  }
}
