\name{binary}
\alias{binary}
\docType{data}
\title{Example data set with a binary endpoint}
\description{
Example data set based on a trial with a binary primary endpoint.
}
\usage{data(binary)}
\format{
  A data set with 100 observations and 2 variables:
  \describe{
    \item{\code{dose}}{Dose level (0 g, 0.05 g, 0.2 g, 0.6 g and 1 g).}
    \item{\code{resp}}{Binary outcome variable (0 or 1). The value of 1 indicates a beneficial outcome.}
    }
}
\keyword{datasets}
