\name{normal}
\alias{normal}
\docType{data}
\title{Example data set with a continuous endpoint}
\description{
Example data set based on a trial with a continuous (normally distributed) primary endpoint.
}
\usage{data(normal)}
\format{
  A data set with 100 observations and 2 variables:
  \describe{
    \item{\code{dose}}{Dose level (0 g, 0.05 g, 0.2 g, 0.6 g and 1 g).}
    \item{\code{resp}}{Continuous outcome variable. A higher value of this variable indicates a beneficial outcome.}
    }
}
\keyword{datasets}
