\name{SimulationApp}
\alias{SimulationApp}
\title{
Creation of a Shiny-based interface to perform an MCPMod-based simulation of dose-finding trial designs
}
\description{
This function creates a Shiny-based graphical user interface to perform a simulation-based evaluation of dose-finding trial designs using MCPMod.
}
\usage{
SimulationApp()
}
\author{
Alex Dmitrienko <admitrienko@medianainc.com>
}
\seealso{
  \code{\link{MCPModSimulation}}, \code{\link{SimulationReport}}
}
